schizophrenia/video_intercept/SimpleH.264HexFrameDecoder/requirements.txt
Pillow==11.3.0
av==15.1.0